Overhaul breadboard to use Arduino Nano.
Turns out the FM18W08 (as opposed to the regular FM1808) is CMOS logic, which means the lower cutoff for a logical 1 is 3.5V. So... for compatibility, a 3.3V logic MCU needs the ability to switch between 5V (for the FM1808) and 3.3V (for the FM18W08). That's a pain - the more elegant solution would be to just use a 5V MCU! So that's what we're doing.