Use a low-side MOSFET for switching on/off the memory chip.
Wow, geez, there are a lot of considerations to be made when cutting power to an IC! This commit implements and explains a whole lot of those.

The main reason for using a MOSFET over a BJT is that BJTs have fairly high voltage drop, which could potentially be troublesome.